requirements.txt
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
diskcache==5.6.3
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
itsdangerous==2.2.0
Jinja2==3.1.4
llama_cpp_python==0.2.89
MarkupSafe==2.1.5
numpy==2.1.0
typing_extensions==4.12.2
Werkzeug==3.0.4
